fix(review): Return `UserProfileModel` from all modification methods Handle image updates through `UserProfiles` class, as they externally get exposed through the same `UserProfileModel` as everything else
