+++ modules/new_serverpod_auth/serverpod_auth_profile/serverpod_auth_profile_server/lib/serverpod_auth_profile_server.dart
@@ -1,5 +1,5 @@
 export 'src/business/user_profile_config.dart';
-export 'src/business/user_profile_images.dart';
 export 'src/business/user_profiles.dart';
 export 'src/generated/endpoints.dart';
 export 'src/generated/protocol.dart' show Protocol, UserProfileModel;
+export 'src/util/default_image_generator.dart';
